Move DatabaseScoreList to other score db models
